+++ src/main/c/xmq.c
@@ -49,6 +49,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include"parts/yaep_vlobject.h"
 #include"parts/yaep.h"
 #include"parts/yaep_structs.h"
+#include"parts/yaep_cspc.h"
 #include"parts/yaep_util.h"
 #include"parts/yaep_symbols.h"
 #include"parts/yaep_terminal_bitset.h"
@@ -4993,6 +4994,7 @@ char *xmqLinePrintf(XMQLineConfig *lc, const char *element_name, ...)
 #include"parts/xmq_parser.c"
 #include"parts/xmq_printer.c"
 #include"parts/yaep_allocate.c"
+#include"parts/yaep_cspc.c"
 #include"parts/yaep_hashtab.c"
 #include"parts/yaep_objstack.c"
 #include"parts/yaep_vlobject.c"
